+++ lib/rails_mermaid_erd/model_loader.rb
@@ -3,6 +3,10 @@
 module RailsMermaidErd
   class ModelLoader
     def load
+      unless defined?(Rails)
+        raise RailsMermaidErd::Error, "Rails is not loaded"
+      end
+      
       Rails.application.eager_load! unless Rails.configuration.cache_classes
       ActiveRecord::Base.descendants.reject(&:abstract_class?)
     end